Refactoring random bits and bobs
